Resolved error in ReplayBuffer that was save actions as longs instead of floats.

+++ code/ReplayBuffer.py
@@ -73,7 +73,7 @@ class ReplayBuffer:
 		'''
 		sampled_experiences = random.sample(self.memory,k=self.batch_size)
 		states = torch.from_numpy(np.vstack([e.state for e in sampled_experiences if e is not None])).float().to(device)
-		actions = torch.from_numpy(np.vstack([e.action for e in sampled_experiences if e is not None])).long().to(device)
+		actions = torch.from_numpy(np.vstack([e.action for e in sampled_experiences if e is not None])).float().to(device)
 		rewards = torch.from_numpy(np.vstack([e.reward for e in sampled_experiences if e is not None])).float().to(device)
 		next_states = torch.from_numpy(np.vstack([e.next_state for e in sampled_experiences if e is not None])).float().to(device)
 		dones = torch.from_numpy(np.vstack([e.done for e in sampled_experiences if e is not None]).astype(np.uint8)).float().to(device)
